Refactor and Fix bug
